+++ .gitignore
@@ -1,7 +1,7 @@
 node_modules
 package-lock.json
 yarn.lock
-pnpm-lock.yaml
+bun.lock
 
 .DS_Store
 .cache
