+++ lib/server/db.ts
@@ -4,6 +4,4 @@ import postgres from "postgres";
 import * as schema from "./schema";
 
 const driver = postgres(process.env.DATABASE_URL as string);
-
 export const db = drizzle({ client: driver, schema });
-export const table = schema;
